+++ server/api/project_admin_api.py
@@ -1,9 +1,7 @@
 from flask_restful import Resource, request, current_app
 from schematics.exceptions import DataError
-from server.models.dtos.message_dto import MessageDTO
 from server.models.dtos.project_dto import DraftProjectDTO, ProjectDTO
 from server.services.authentication_service import token_auth, tm
-from server.services.message_service import MessageService
 from server.services.project_admin_service import ProjectAdminService, InvalidGeoJson, InvalidData, \
     ProjectAdminServiceError, NotFound
 from server.services.validator_service import ValidatorService
@@ -473,70 +471,6 @@ class ProjectsForAdminAPI(Resource):
             return admin_projects.to_primitive(), 200
         except NotFound:
             return {"Error": "No comments found"}, 404
-        except Exception as e:
-            error_msg = f'Project GET - unhandled error: {str(e)}'
-            current_app.logger.critical(error_msg)
-            return {"error": error_msg}, 500
-
-
-class ProjectsMessageAll(Resource):
-
-    @tm.pm_only()
-    @token_auth.login_required
-    def post(self, project_id):
-        """
-        Send message to all contributors to a project
-        ---
-        tags:
-            - project-admin
-        produces:
-            - application/json
-        parameters:
-            - in: header
-              name: Authorization
-              description: Base64 encoded session token
-              required: true
-              type: string
-              default: Token sessionTokenHere==
-            - name: project_id
-              in: path
-              description: The unique project ID
-              required: true
-              type: integer
-              default: 1
-            - in: body
-              name: body
-              required: true
-              description: JSON object for creating draft project
-              schema:
-                  properties:
-                      subject:
-                          type: string
-                          default: Thanks 
-                          required: true
-                      message:
-                          type: string
-                          default: Thanks for your contribution
-                          required: true
-        responses:
-            200:
-                description: All mapped tasks validated
-            401:
-                description: Unauthorized - Invalid credentials
-            500:
-                description: Internal Server Error
-        """
-        try:
-            message_dto = MessageDTO(request.get_json())
-            message_dto.from_user = tm.authenticated_user_id
-            message_dto.validate()
-        except DataError as e:
-            current_app.logger.error(f'Error validating request: {str(e)}')
-            return str(e), 400
-
-        try:
-            MessageService.send_message_to_all_contributors(project_id, message_dto)
-            return {"Success": "Messages sent"}, 200
         except Exception as e:
             error_msg = f'Project GET - unhandled error: {str(e)}'
             current_app.logger.critical(error_msg)
